+++ pdf-gen/src/types/index.ts
@@ -1,0 +1,10 @@
+export interface Form {
+    id: number | string;
+    name: string;
+    fields: Field[]
+}
+  
+export interface Field {
+    id: number | string;
+    name: string;
+}
